core/src/main/java/org/geysermc/floodgate/core/crypto/aes/AesKeyCodec.java
/*
 * Copyright (c) 2019-2023 GeyserMC. http://geysermc.org
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 * THE SOFTWARE.
 *
 * @author GeyserMC
 * @link https://github.com/GeyserMC/Floodgate
 */

package org.geysermc.floodgate.core.crypto.aes;

import java.util.Base64;
import javax.crypto.SecretKey;
import javax.crypto.spec.SecretKeySpec;
import org.geysermc.floodgate.core.crypto.KeyCodecSingle;

public final class AesKeyCodec implements KeyCodecSingle {
    @Override
    public SecretKey decode(byte[] keyBytes) {
        // we only started encoding keys with base64 since 3.0
        // we ignore if it's not valid base64 to allow those old keys to keep working
        try {
            keyBytes = Base64.getDecoder().decode(keyBytes);
        } catch (IllegalArgumentException ignored) {}

        return new SecretKeySpec(keyBytes, "AES");
    }

    @Override
    public byte[] encode(SecretKey key) {
        return Base64.getEncoder().encode(key.getEncoded());
    }
}
